Delete slurm interface
